+++ examples/runtimes/main.tf
@@ -93,6 +93,35 @@ module "java21_lambda_function" {
   ]
 }
 
+module "dotnet8_lambda_function" {
+  source = "../../"
+
+  function_name = "${random_pet.this.id}-dotnet8"
+
+  attach_cloudwatch_logs_policy     = false
+  cloudwatch_logs_retention_in_days = 1
+
+  create_lambda_function_url = true
+
+  handler       = "bootstrap"
+  runtime       = "dotnet8"
+  architectures = ["arm64"] # x86_64 (--runtime linux-x64) or arm64 (--runtime linux-arm64)
+  timeout       = 30
+
+  trigger_on_package_timestamp = false
+
+  source_path = [
+    {
+      path = "${path.module}/../fixtures/runtimes/dotnet8"
+      commands = [
+        "dotnet publish --framework net8.0 --configuration Relesase --runtime linux-arm64 --output ./publish",
+        "cd publish",
+        ":zip",
+      ]
+    }
+  ]
+}
+
 resource "random_pet" "this" {
   length = 2
 }