+++ wrappers/docker-build/versions.tf
@@ -1,14 +1,14 @@
 terraform {
-  required_version = ">= 1.0"
+  required_version = ">= 1.5.7"
 
   required_providers {
     aws = {
       source  = "hashicorp/aws"
-      version = ">= 4.22"
+      version = ">= 6.0"
     }
     docker = {
       source  = "kreuzwerker/docker"
-      version = ">= 3.0"
+      version = ">= 3.5.0"
     }
     null = {
       source  = "hashicorp/null"